Rename api_table component to component_api
