+++ kolibri/plugins/style_guide/assets/src/views/content/component_api.vue
