+++ kolibri/plugins/style_guide/assets/src/views/content/components/buttons.vue
@@ -140,7 +140,7 @@
     <vuep class="code-examples" :template="codeExamplesTemplate"></vuep>
 
     <h2>API</h2>
-    <api-table :api="api"></api-table>
+    <component-api :api="api"></component-api>
   </div>
 
 </template>
@@ -208,7 +208,7 @@
 
   module.exports = {
     components: {
-      'api-table': require('../api_table'),
+      'component-api': require('../component_api'),
     },
     data: () => ({
       codeExamplesTemplate,
