+++ kolibri/plugins/style_guide/assets/src/views/content/components/radio_buttons.vue
@@ -69,7 +69,7 @@
     <vuep class="code-examples" :template="codeExamplesTemplate"></vuep>
 
     <h2>API</h2>
-    <api-table :api="api"></api-table>
+    <component-api :api="api"></component-api>
 
   </div>
 
@@ -130,7 +130,7 @@
 
   module.exports = {
     components: {
-      'api-table': require('../api_table'),
+      'component-api': require('../component_api'),
     },
     data: () => ({
       codeExamplesTemplate,
