+++ kolibri/plugins/style_guide/assets/src/views/content/components/text_fields.vue
@@ -7,7 +7,7 @@
     <vuep :template="codeExamplesTemplate"></vuep>
 
     <h2>API</h2>
-    <api-table :api="api"></api-table>
+    <component-api :api="api"></component-api>
   </div>
 
 </template>
@@ -44,7 +44,7 @@
 
   module.exports = {
     components: {
-      'api-table': require('../api_table'),
+      'component-api': require('../component_api'),
     },
     data: () => ({
       codeExamplesTemplate,
